requirements.txt
asgiref==3.10.0
Django==5.2.7
gunicorn==23.0.0
numpy==2.2.6
packaging==25.0
pillow==11.3.0
sqlparse==0.5.3
tzdata==2025.2
